requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
attrs==25.3.0
Authlib==1.6.0
backoff==2.2.1
bleach==6.2.0
caldav==2.0.1
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
coverage==7.9.2
cryptography==45.0.5
cyclopts==3.22.2
dataclasses-json==0.6.7
distro==1.9.0
Django==5.2.4
django-encrypted-model-fields==0.6.5
djangorestframework==3.16.0
dnspython==2.7.0
docstring_parser==0.16
docutils==0.21.2
dotenv==0.9.9
email_validator==2.2.0
exceptiongroup==1.3.0
fastmcp==2.10.5
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.5.1
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.4
icalendar==6.3.1
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-mistralai==0.2.11
langchain-ollama==0.3.4
langchain-openai==0.3.28
langchain-text-splitters==0.3.8
langfuse==3.2.0
langgraph==0.5.2
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.72
langsmith==0.4.5
lxml==6.0.0
Markdown==3.8
markdown-it-py==3.0.0
marshmallow>=3.26.1
mcp==1.11.0
mdurl==0.1.2
multidict==6.6.3
mypy_extensions==1.1.0
numpy==2.3.1
ollama==0.5.1
openai==1.95.1
openapi-pydantic==0.5.1
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-exporter-otlp-proto-http==1.35.0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pluggy==1.6.0
propcache==0.3.2
protobuf==6.31.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyperclip==1.9.0
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
recurring-ical-events==3.8.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rich-rst==1.3.1
rpds-py==0.26.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
sse-starlette==2.4.1
starlette==0.47.1
tdqm==0.0.1
tenacity==9.1.2
tiktoken==0.9.0
tokenizer==3.4.5
tokenizers==0.21.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
vobject==0.9.9
webencodings==0.5.1
wrapt==1.17.2
x-wr-timezone==2.0.1
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
gunicorn==20.1.0
psycopg2-binary==2.9.9
setuptools<81